+++ jgclark.DashboardReact/src/react/components/AppContext.jsx
@@ -3,17 +3,16 @@
  ****************************************************************************************************************************
  * This is a shared context provider for NotePlan React Apps. It provides a context for the app to communicate with the plugin.
  * It also provides a context for the plugin to communicate with the app.
- * Plus local storage of reactSettings
  * @usage import { useAppContext } from './AppContext.jsx'
- * @usage const {sendActionToPlugin, sendToPlugin, dispatch, TPluginData, reactSettings, setReactSettings}  = useAppContext()
+ * @usage const {sendActionToPlugin, sendToPlugin, dispatch, TPluginData, reactSettings, setReactSettings, updatePluginData, sharedSettings, setSharedSettings}  = useAppContext()
  *
  ****************************************************************************************************************************/
 // @flow
 
 import React, { createContext, useContext, useEffect, type Node } from 'react'
-import { type TReactSettings, type TPluginData } from '../../types'
+import { type TReactSettings, type TPluginData, type TSharedSettings } from '../../types'
 import { logDebug } from '@helpers/react/reactDev'
-logDebug(`AppContext`, `outside component code`)
+
 
 /****************************************************************************************************************************
  *                             TYPES
@@ -27,18 +26,13 @@ export type AppContextType = {
   reactSettings: ?TReactSettings,
   setReactSettings: (any) => void,
   updatePluginData: (newData: TPluginData, messageForLog?: string) => void,
+  sharedSettings: TSharedSettings,
+  setSharedSettings: (any) => void
 }
 
 type Props = {
   children?: Node,
-  sendActionToPlugin: (command: string, dataToSend: any, additionalInfo?: string, updateGlobalData?: boolean) => void,
-  sendToPlugin: ([string, any, string]) => void,
-  dispatch: (command: string, dataToSend: any, messageForLog?: string) => void,
-  pluginData: TPluginData,
-  reactSettings: TReactSettings,
-  setReactSettings: (newSettings: TReactSettings, msgForLog?: string) => void,
-  updatePluginData: (newData: TPluginData, messageForLog?: string) => void,
-}
+} & AppContextType;
 
 // Default context value with initial reactSettings and functions.
 const defaultContextValue: AppContextType = {
@@ -49,6 +43,8 @@ const defaultContextValue: AppContextType = {
   reactSettings: {}, // Initial empty reactSettings local
   setReactSettings: () => {},
   updatePluginData: () => {}, // Placeholder function, actual implementation below.
+  sharedSettings: {},
+  setSharedSettings: () => {},
 }
 
 /****************************************************************************************************************************
@@ -61,7 +57,8 @@ const AppContext = createContext<AppContextType>(defaultContextValue)
  *                             CONTEXT PROVIDER FUNCTIONS
  ****************************************************************************************************************************/
 
-export const AppProvider = ({ children, sendActionToPlugin, sendToPlugin, dispatch, pluginData, reactSettings, setReactSettings, updatePluginData }: Props): Node => {
+// eslint-disable-next-line max-len
+export const AppProvider = ({ children, sendActionToPlugin, sendToPlugin, dispatch, pluginData, reactSettings, setReactSettings, updatePluginData, sharedSettings, setSharedSettings }: Props): Node => {
   logDebug(`AppProvider`, `inside component code`)
 
   const contextValue: AppContextType = {
@@ -72,6 +69,8 @@ export const AppProvider = ({ children, sendActionToPlugin, sendToPlugin, dispat
     reactSettings,
     setReactSettings,
     updatePluginData,
+    sharedSettings,
+    setSharedSettings,
   }
 
   useEffect(() => {